Review comment from Stefan Hoehn.
- Example in README
- Renames to make the Panels/Shapes easier to understand
