+++ bundles/org.openhab.binding.nanoleaf/README.md
@@ -104,7 +104,15 @@ Compare the following output with the right picture at the beginning of the arti
                                     41451                                     
 
 ```
-         
+
+## State
+
+The state channel shows an image of the panels on the wall. You have to configure things for each panel to
+get the correct color. Since the colors of the panels can make it difficult to see the panel ids, please use
+the layout channel where the background color is always white to identify them.
+
+![Image](doc/NanoCanvas_rendered.jpg)
+
 ## Thing Configuration
 
 The controller thing has the following parameters:
