+++ bundles/org.openhab.binding.nanoleaf/src/main/java/org/openhab/binding/nanoleaf/internal/layout/shape/Shape.java
@@ -12,7 +12,9 @@
  */
 package org.openhab.binding.nanoleaf.internal.layout.shape;
 
+import java.awt.Color;
 import java.awt.Graphics2D;
+import java.awt.Polygon;
 import java.util.List;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
@@ -21,59 +23,74 @@ import org.openhab.binding.nanoleaf.internal.layout.ImagePoint2D;
 import org.openhab.binding.nanoleaf.internal.layout.PanelState;
 import org.openhab.binding.nanoleaf.internal.layout.Point2D;
 import org.openhab.binding.nanoleaf.internal.layout.ShapeType;
+import org.openhab.core.library.types.HSBType;
 
 /**
- * Shape that can be drawn.
+ * Draws shapes is panel with a single LED.
  *
  * @author Jørgen Austvik - Initial contribution
  */
 @NonNullByDefault
-public abstract class Shape {
-    private final ShapeType shapeType;
+public abstract class Shape extends Panel {
 
-    public Shape(ShapeType shapeType) {
-        this.shapeType = shapeType;
+    private final Point2D position;
+    private final int orientation;
+    private final int panelId;
+
+    public Shape(ShapeType shapeType, int panelId, Point2D position, int orientation) {
+        super(shapeType);
+        this.position = position;
+        this.orientation = orientation;
+        this.panelId = panelId;
     }
 
-    public ShapeType getShapeType() {
-        return shapeType;
+    public Point2D getPosition() {
+        return position;
     }
 
-    /**
-     * Calculates the minimal bounding rectangle around an outline.
-     *
-     * @param outline The outline to find the minimal bounding rectangle around
-     * @return The opposite points of the minimum bounding rectangle around this shape.
-     */
-    public Point2D[] findBounds(List<ImagePoint2D> outline) {
-        int minX = Integer.MAX_VALUE;
-        int minY = Integer.MAX_VALUE;
-        int maxX = Integer.MIN_VALUE;
-        int maxY = Integer.MIN_VALUE;
+    public int getOrientation() {
+        return orientation;
+    };
 
-        for (ImagePoint2D point : outline) {
-            maxX = Math.max(point.getX(), maxX);
-            maxY = Math.max(point.getY(), maxY);
-            minX = Math.min(point.getX(), minX);
-            minY = Math.min(point.getY(), minY);
-        }
-
-        return new Point2D[] { new Point2D(minX, minY), new Point2D(maxX, maxY) };
+    protected int getPanelId() {
+        return panelId;
     }
 
-    /**
-     * Generate the outline of the shape.
-     *
-     * @return The points that make up this shape.
-     */
+    @Override
     public abstract List<Point2D> generateOutline();
 
     /**
-     * Draws the shape on the the supplied graphics.
-     *
      * @param graphics The picture to draw on
-     * @param settings Information on how to draw
-     * @param state The state of the panels to draw
+     * @param outline Outline of the shape to draw inside
+     * @return The position where the label of the shape should be placed
      */
-    public abstract void draw(Graphics2D graphics, DrawingSettings settings, PanelState state);
+    protected abstract ImagePoint2D labelPosition(Graphics2D graphics, List<ImagePoint2D> outline);
+
+    @Override
+    public void draw(Graphics2D graphics, DrawingSettings settings, PanelState state) {
+        List<ImagePoint2D> outline = settings.generateImagePoints(generateOutline());
+
+        Polygon p = new Polygon();
+        for (int i = 0; i < outline.size(); i++) {
+            ImagePoint2D pos = outline.get(i);
+            p.addPoint(pos.getX(), pos.getY());
+        }
+
+        HSBType color = state.getHSBForPanel(getPanelId());
+        graphics.setColor(new Color(color.getRGB()));
+        if (settings.shouldFillWithColor()) {
+            graphics.fillPolygon(p);
+        }
+
+        if (settings.shouldDrawOutline()) {
+            graphics.setColor(settings.getOutlineColor());
+            graphics.drawPolygon(p);
+        }
+
+        if (settings.shouldDrawLabels()) {
+            graphics.setColor(settings.getLabelColor());
+            ImagePoint2D textPos = labelPosition(graphics, outline);
+            graphics.drawString(Integer.toString(getPanelId()), textPos.getX(), textPos.getY());
+        }
+    }
 }
